ip_tracking/requirements.txt
Django>=4.0,<5.0
# django-ipgeolocation
django-ratelimit
celery
django-celery-beat
djangorestframework
drf-yasg
mysqlclient
